requirements.txt
langgraph
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-openai
notebook
tavily-python
wikipedia
trustcall
langgraph-cli[inmem]
langchain
langchain_gigachat
langchain_chroma
rich
python-dotenv
nest_asyncio
pyppeteer